Crate node_primitives

Low-level types used throughout the Substrate code.

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex
The type for looking up accounts. We don't expect more than 4 billion of them, but you never know...

AuthorityId
The Ed25519 pub key of an session that belongs to an authority of the chain. This is exactly equivalent to what the substrate calls an "authority".

AuthoritySignature
Alias to 512-bit hash when used in the context of a session signature on the chain.

Balance
Balance of an account.

Block
Block type.

BlockId
Block ID.

BlockNumber
An index to a block.

Hash
A hash of some data used by the chain.

Header
Header type.

Index
Index of a transaction in the chain.

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.

Timestamp
A timestamp: seconds since the unix epoch.

UncheckedExtrinsic
Opaque, encoded, unchecked extrinsic.